+++ src/nationalarchives/components/footer/footer.stories.js
@@ -81,6 +81,7 @@ Standard.args = {
   navigation: [
     {
       title: "About us",
+      hideTitle: true,
       items: [
         {
           text: "Our role",